extract `setUserSetting` & improve tests
