+++ client/state/user-settings/thunks/index.js
@@ -1,2 +1,3 @@
 export { default as saveTwoStepSMSSettings } from './save-two-step-sms-settings';
 export { default as saveUnsavedUserSettings } from './save-unsaved-user-settings';
+export { default as setUserSetting } from './set-user-setting';
